refactor(core): move telemetry to sec layers
[no changelog]
