+++ core/embed/sec/telemetry/inc/sec/telemetry.h
